Non-recursive search, C implementation
I was curious what kind of speedup I'd get making DancingLinks.search iterative
and not recurse.  Looks to be minimal, 6% faster or so non-recursive.

I was also curious how much faster a C implementation would be, so I ported the
Java version over.  It looks to be about 40% faster than the Java version.